+++ GRDB/Utils/Utils.swift
@@ -43,7 +43,6 @@ extension Optional : _OptionalProtocol {
 
 // MARK: - Internal
 
-/// Reserved for GRDB: do not use.
 func GRDBPrecondition(_ condition: @autoclosure() -> Bool, _ message: @autoclosure() -> String = "", file: StaticString = #file, line: UInt = #line) {
     /// Custom precondition function which aims at solving
     /// https://bugs.swift.org/browse/SR-905 and
@@ -53,6 +52,15 @@ func GRDBPrecondition(_ condition: @autoclosure() -> Bool, _ message: @autoclosu
     if !condition() {
         fatalError(message, file: file, line: line)
     }
+}
+
+func GRDBPreconditionFailure(_ message: @autoclosure() -> String = "", file: StaticString = #file, line: UInt = #line) -> Never {
+    /// Custom precondition function which aims at solving
+    /// https://bugs.swift.org/browse/SR-905 and
+    /// https://github.com/groue/GRDB.swift/issues/37
+    ///
+    /// TODO: remove this function when https://bugs.swift.org/browse/SR-905 is solved.
+    fatalError(message, file: file, line: line)
 }
 
 // Workaround Swift inconvenience around factory methods of non-final classes